+++ collector/src/main/java/com/bitmovin/analytics/stateMachines/PlayerState.java
@@ -19,10 +19,12 @@ public enum PlayerState {
     BUFFERING {
         @Override
         void onEnterState(PlayerStateMachine machine) {
+            machine.enableRebufferHeartbeat();
         }
 
         @Override
         void onExitState(PlayerStateMachine machine, long elapsedTime, PlayerState desintationPlayerState) {
+            machine.disableRebufferHeartbeat();
             for (StateMachineListener listener : machine.getListeners()) {
                 long elapsedTimeOnEnter = machine.getElapsedTimeOnEnter();
                 listener.onRebuffering(elapsedTime - elapsedTimeOnEnter);