+++ collector/src/main/java/com/bitmovin/analytics/stateMachines/PlayerStateMachine.java
@@ -8,7 +8,9 @@ import com.bitmovin.analytics.BitmovinAnalyticsConfig;
 import com.bitmovin.analytics.data.ErrorCode;
 import com.bitmovin.analytics.utils.Util;
 
+import java.lang.reflect.Array;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 public class PlayerStateMachine {
@@ -25,6 +27,8 @@ public class PlayerStateMachine {
     private ErrorCode errorCode;
     private String impressionId;
     private Handler heartbeatHandler = new Handler();
+    private int currentRebufferingIntervalIndex = 0;
+    private static List<Integer> rebufferingIntervals = Arrays.asList(3000, 5000, 10000, 30000, 59700);
     private int heartbeatDelay = 59700; // default to 60 seconds
     private final BitmovinAnalytics analytics;
 
@@ -38,14 +42,7 @@ public class PlayerStateMachine {
     public void enableHeartbeat() {
         heartbeatHandler.postDelayed(new Runnable() {
             public void run() {
-                long elapsedTime = Util.getElapsedTime();
-                long elapsedTimeOnEnter = getElapsedTimeOnEnter();
-                videoTimeEnd = analytics.getPosition();
-                for (StateMachineListener listener : getListeners()) {
-                    listener.onHeartbeat(elapsedTime - elapsedTimeOnEnter);
-                }
-                PlayerStateMachine.this.elapsedTimeOnEnter = elapsedTime;
-                videoTimeStart = videoTimeEnd;
+                triggerHeartbeat();
                 heartbeatHandler.postDelayed(this, heartbeatDelay);
             }
         }, heartbeatDelay);
@@ -53,6 +50,32 @@ public class PlayerStateMachine {
 
     public void disableHeartbeat() {
         heartbeatHandler.removeCallbacksAndMessages(null);
+    }
+
+    public void enableRebufferHeartbeat() {
+        heartbeatHandler.postDelayed(new Runnable() {
+            public void run() {
+                triggerHeartbeat();
+                currentRebufferingIntervalIndex = Math.min(currentRebufferingIntervalIndex + 1, rebufferingIntervals.size() - 1);
+                heartbeatHandler.postDelayed(this, rebufferingIntervals.get(currentRebufferingIntervalIndex));
+            }
+        }, rebufferingIntervals.get(currentRebufferingIntervalIndex));
+    }
+
+    public void disableRebufferHeartbeat() {
+        currentRebufferingIntervalIndex = 0;
+        heartbeatHandler.removeCallbacksAndMessages(null);
+    }
+
+    private void triggerHeartbeat() {
+        long elapsedTime = Util.getElapsedTime();
+        long elapsedTimeOnEnter = getElapsedTimeOnEnter();
+        videoTimeEnd = analytics.getPosition();
+        for (StateMachineListener listener : getListeners()) {
+            listener.onHeartbeat(elapsedTime - elapsedTimeOnEnter);
+        }
+        PlayerStateMachine.this.elapsedTimeOnEnter = elapsedTime;
+        videoTimeStart = videoTimeEnd;
     }
 
     public void resetStateMachine() {