+++ collector/src/test/java/com/bitmovin/analytics/utils/DownloadSpeedMeterTest.kt
@@ -56,7 +56,4 @@ class DownloadSpeedMeterTest {
         Assertions.assertThat(info.avgDownloadSpeed).isEqualTo(504.0f);
     }
 
-
-
-
 }